third_party/blink/web_tests/platform/mac/external/wpt/html/canvas/offscreen/text/2d.text.drawing.style.wordSpacing.measure-expected.txt

This is a testharness.js-based test.
FAIL Testing word spacing with different length units assert_approx_equals: word spacing doesn't work. expected 90.58702256944444 +/- 0.1 but got 81.5283203125
Harness: the test ran to completion.
